packages/report/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "wealth_management_portal_report" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --ignore=tests/integration
 --cov
 --cov-report html:'../../coverage/packages/report/html'
 --cov-report xml:'../../coverage/packages/report/coverage.xml'
 --html='../../reports/packages/report/unittests/html/index.html'
 --junitxml='../../reports/packages/report/unittests/junit.xml'
"""
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
]

[tool.hatch.build.targets.wheel]
packages = [ "wealth_management_portal_report" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 120

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "wealth_management_portal.report"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "wealth_management_portal.common_auth",
  "bedrock-agentcore==0.1.7",
  "fastapi==0.128.5",
  "boto3==1.42.44",
  "mcp==1.26.0",
  "strands-agents==1.25.0",
  "uvicorn==0.40.0",
  "pydantic>=2.12.5",
  "jinja2>=3.1.0",
  "matplotlib>=3.8.0",
  "markdown>=3.5.0",
  "weasyprint>=60.0"
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[tool.uv.sources]
"wealth_management_portal.common_auth" = { workspace = true }

[dependency-groups]
dev = [
 "fastapi[standard]==0.128.5",
 "python-dotenv>=1.2.1",
 "moto[s3]>=5.0.0",
]
